do not duplicate thesaurus with diferent casing
