+++ app/api/csv/specs/typeParsers.spec.js
@@ -1,10 +1,9 @@
 import moment from 'moment';
 
 import entities, { model } from 'api/entities';
-import thesauris from 'api/thesauris';
 import db from 'api/utils/testing_db';
 
-import fixtures, { thesauri1Id, templateToRelateId } from './fixtures';
+import fixtures, { templateToRelateId } from './fixtures';
 import typeParsers from '../typeParsers';
 
 describe('csvLoader typeParsers', () => {
@@ -73,105 +72,6 @@ describe('csvLoader typeParsers', () => {
 
       expected = await typeParsers.date({ date_prop: '1/1/1996 23:59:59' }, templateProp);
       expect(moment.utc(expected, 'X').format('DD-MM-YYYY')).toEqual('01-01-1996');
-    });
-  });
-
-  describe('select', () => {
-    it('should create thesauri value and return the id', async () => {
-      const templateProp = { name: 'select_prop', content: thesauri1Id };
-
-      const value1 = await typeParsers.select({ select_prop: 'value' }, templateProp);
-      const value2 = await typeParsers.select({ select_prop: 'value2' }, templateProp);
-      const thesauri1 = await thesauris.getById(thesauri1Id);
-
-      expect(thesauri1.values[1].label).toBe('value');
-      expect(thesauri1.values[2].label).toBe('value2');
-      expect(value1).toBe(thesauri1.values[1].id);
-      expect(value2).toBe(thesauri1.values[2].id);
-    });
-
-    it('should not create repeated values', async () => {
-      const templateProp = { name: 'select_prop', content: thesauri1Id };
-
-      await typeParsers.select({ select_prop: 'value4' }, templateProp);
-      await typeParsers.select({ select_prop: 'value ' }, templateProp);
-      await typeParsers.select({ select_prop: 'value' }, templateProp);
-      await typeParsers.select({ select_prop: ' value ' }, templateProp);
-      await typeParsers.select({ select_prop: 'value4' }, templateProp);
-      const thesauri1 = await thesauris.getById(thesauri1Id);
-
-      expect(thesauri1.values.length).toBe(2);
-    });
-
-    it('should not create blank values', async () => {
-      const templateProp = { name: 'select_prop', content: thesauri1Id };
-      const rawEntity = { select_prop: ' ' };
-
-      const value = await typeParsers.select(rawEntity, templateProp);
-      const thesauri1 = await thesauris.getById(thesauri1Id);
-
-      expect(thesauri1.values.length).toBe(1);
-      expect(value).toBe(null);
-    });
-  });
-
-  describe('multiselect', () => {
-    let value1;
-    let value2;
-    let value3;
-    let value4;
-    let thesauri1;
-
-    const templateProp = { name: 'multiselect_prop', content: thesauri1Id };
-
-    beforeAll(async () => {
-      value1 = await typeParsers.multiselect(
-        { multiselect_prop: 'value4' },
-        templateProp
-      );
-
-      value2 = await typeParsers.multiselect(
-        { multiselect_prop: 'value1|value3| value3' },
-        templateProp
-      );
-
-      value3 = await typeParsers.multiselect(
-        { multiselect_prop: 'value1| value2' },
-        templateProp
-      );
-
-      value4 = await typeParsers.multiselect(
-        { multiselect_prop: 'value1|value2' },
-        templateProp
-      );
-
-      await typeParsers.multiselect(
-        { multiselect_prop: '' },
-        templateProp
-      );
-
-      await typeParsers.multiselect(
-        { multiselect_prop: '|' },
-        templateProp
-      );
-
-      thesauri1 = await thesauris.getById(thesauri1Id);
-    });
-
-    it('should create thesauri values and return an array of ids', async () => {
-      expect(thesauri1.values[0].label).toBe(' value4 ');
-      expect(thesauri1.values[1].label).toBe('value1');
-      expect(thesauri1.values[2].label).toBe('value3');
-      expect(thesauri1.values[3].label).toBe('value2');
-
-      expect(value1).toEqual([thesauri1.values[0].id]);
-      expect(value2).toEqual([thesauri1.values[1].id, thesauri1.values[2].id]);
-      expect(value3).toEqual([thesauri1.values[1].id, thesauri1.values[3].id]);
-      expect(value4).toEqual([thesauri1.values[1].id, thesauri1.values[3].id]);
-    });
-
-    it('should not create blank values, or repeat values', async () => {
-      expect(thesauri1.values.length).toBe(4);
     });
   });
