+++ app/api/csv/typeParsers.js
@@ -1,12 +1,19 @@
 import entities from 'api/entities';
-import thesauris from 'api/thesauris';
 import url from 'url';
 
 import { unique, emptyString } from 'api/utils/filters';
+
 import geolocation from './typeParsers/geolocation.js';
+import multiselect from './typeParsers/multiselect.js';
+import select from './typeParsers/select.js';
 
 export default {
   geolocation,
+
+  select,
+
+  multiselect,
+
   async default(entityToImport, templateProperty) {
     return this.text(entityToImport, templateProperty);
   },
@@ -64,59 +71,5 @@ export default {
 
     const toRelateEntities = await entities.get({ title: { $in: values.map(v => new RegExp(`\\s?${v}\\s?`, 'i')) } });
     return toRelateEntities.map(e => e.sharedId).filter((id, index, ids) => ids.indexOf(id) === index);
-  },
-
-  async multiselect(entityToImport, templateProperty) {
-    const currentThesauri = await thesauris.getById(templateProperty.content);
-
-    const values = entityToImport[templateProperty.name].split('|')
-    .map(v => v.trim())
-    .filter(emptyString)
-    .filter(unique);
-
-    const newValues = values.filter(v =>
-      !currentThesauri.values.find(cv => cv.label.trim() === v));
-
-    if (!newValues.length) {
-      return currentThesauri.values
-      .filter(value => values.includes(value.label.trim()))
-      .map(value => value.id);
-    }
-
-    const updated = await thesauris.save({
-      ...currentThesauri,
-      values: currentThesauri.values.concat(
-        newValues.map(label => ({ label }))
-      )
-    });
-
-    return updated.values
-    .filter(value => values.includes(value.label))
-    .map(value => value.id);
-  },
-
-  async select(entityToImport, templateProperty) {
-    const currentThesauri = await thesauris.getById(templateProperty.content);
-    if (entityToImport[templateProperty.name].trim() === '') {
-      return null;
-    }
-
-    const thesauriMatching =
-      v => v.label.trim() === entityToImport[templateProperty.name].trim();
-
-    const value = currentThesauri.values.find(thesauriMatching);
-
-    if (value) {
-      return value.id;
-    }
-
-    const updated = await thesauris.save({
-      ...currentThesauri,
-      values: currentThesauri.values.concat([{
-        label: entityToImport[templateProperty.name]
-      }])
-    });
-
-    return updated.values.find(thesauriMatching).id;
   },
 };
